refactor: extract logger types to logger-types.ts
- Add src/utils/logger-types.ts with logger type definitions and LOG_LEVELS
- Remove duplicated type declarations from src/utils/logger.ts and import types/LOG_LEVELS
- Update src/utils/README.md to document logger.ts and logger-types.ts
- No runtime behavior changes; refactor for clearer separation of types and implementation
